Auth and multitenancy

+++ app/todo-app/todo-endpoints.go
@@ -2,6 +2,7 @@ package todo
 
 import (
 	"fmt"
+	"todo_app/app"
 	"todo_app/common"
 	"todo_app/store"
 
@@ -9,37 +10,28 @@ import (
 	"github.com/google/uuid"
 )
 
-type App struct {
-	*fiber.App
-}
-
-func New() *App {
-	return &App{
-		App: fiber.New(),
-	}
-}
-
-func (app *App) UseTodoEndpoints(store *store.Store) {
-	todoGroup := app.Group("api/v1/todo")
+func UseEndpoints(application *app.App, store *store.Store) {
+	todoGroup := application.Group("api/v1/todo")
 	{
-		todoGroup.Get("/", GetAllItems(store))
-		todoGroup.Post("/", AddNewItem(store))
-		todoGroup.Put("/", UpdateItem(store))
-		todoGroup.Delete("/:id", DeleteItem(store))
+		todoGroup.Get("/", getAllItems(store))
+		todoGroup.Post("/", addNewItem(store))
+		todoGroup.Put("/", updateItem(store))
+		todoGroup.Delete("/:id", deleteItem(store))
 	}
 }
 
-func GetAllItems(store *store.Store) func(c *fiber.Ctx) error {
+func getAllItems(store *store.Store) func(c *fiber.Ctx) error {
 	return func(c *fiber.Ctx) error {
+		id := common.GetUserFromContext(c)
 		var todoItems []*TodoItem
-		rows, err := store.Query("select * from TodoItems")
+		rows, err := store.Query("select * from TodoItems where creatorId = $1", id)
 		if err != nil {
 			return err
 		}
 
 		for rows.Next() {
 			var i TodoItem
-			err = rows.Scan(&i.Id, &i.Description, &i.CreatedAt, &i.IsDone)
+			err = rows.Scan(&i.Id, &i.Description, &i.CreatedAt, &i.IsDone, &i.CreatorId)
 			if err != nil {
 				return err
 			}
@@ -55,30 +47,32 @@ func GetAllItems(store *store.Store) func(c *fiber.Ctx) error {
 	}
 }
 
-func AddNewItem(store *store.Store) func(c *fiber.Ctx) error {
+func addNewItem(store *store.Store) func(c *fiber.Ctx) error {
 	return func(c *fiber.Ctx) error {
+		id := common.GetUserFromContext(c)
 		i, err := common.ReadJson[TodoItem](c.Body())
 		if err != nil {
 			return err
 		}
 
-		row, err := store.QueryRow("insert into TodoItems (description, isDone) values ($1, $2) returning id;", i.Description, i.IsDone)
+		row, err := store.QueryRow("insert into TodoItems (creatorId, description, isDone) values ($1, $2, $3) returning id;", id, i.Description, i.IsDone)
 		if err != nil {
 			return err
 		}
 
-		var id uuid.UUID
-		err = row.Scan(&id)
+		var itemId string
+		err = row.Scan(&itemId)
 		if err != nil {
 			return err
 		}
 
-		return c.SendString(id.String())
+		return c.SendString(itemId)
 	}
 }
 
-func UpdateItem(store *store.Store) func(c *fiber.Ctx) error {
+func updateItem(store *store.Store) func(c *fiber.Ctx) error {
 	return func(c *fiber.Ctx) error {
+		id := common.GetUserFromContext(c)
 		updatedItem, err := common.ReadJson[TodoItem](c.Body())
 		if err != nil {
 			c.Status(400)
@@ -86,8 +80,9 @@ func UpdateItem(store *store.Store) func(c *fiber.Ctx) error {
 		}
 
 		row, err := store.QueryRow(
-			"select exists(select 1 from TodoItems where id=$1)",
-			updatedItem.Id)
+			"select exists(select 1 from TodoItems where id=$1 and creatorId = $2)",
+			updatedItem.Id,
+			id)
 		if err != nil {
 			return err
 		}
@@ -111,16 +106,17 @@ func UpdateItem(store *store.Store) func(c *fiber.Ctx) error {
 	}
 }
 
-func DeleteItem(store *store.Store) func(c *fiber.Ctx) error {
+func deleteItem(store *store.Store) func(c *fiber.Ctx) error {
 	return func(c *fiber.Ctx) error {
-		var id uuid.UUID
-		id, err := uuid.Parse(c.Params("id"))
+		id := common.GetUserFromContext(c)
+		var itemId uuid.UUID
+		itemId, err := uuid.Parse(c.Params("id"))
 		if err != nil {
 			c.Status(400)
 			return c.SendString(fmt.Sprintf("Error occured: %s", err))
 		}
 
-		err = store.Execute("delete from TodoItems where id = $1", id)
+		err = store.Execute("delete from TodoItems where id = $1 and creatorId = $2", itemId, id)
 		if err != nil {
 			return err
 		}
